# perf: otimizar font loading para reduzir LCP (T1 - Cirurgia LCP)