--- a/src/app/layout.tsx
+++ b/src/app/layout.tsx
@@ -20,6 +20,11 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="pt">
+      <head>
+        {/* Otimização LCP: Preconnect para Google Fonts */}
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+      </head>
       <body className={`${fontSans.variable} ${fontHeading.variable} antialiased`}>
         <Header />
         {children}