--- a/src/components/Testimonials.tsx
+++ b/src/components/Testimonials.tsx
@@ -1,9 +1,11 @@
 import { t } from "@/lib/i18n";
+
 export default function Testimonials() {
   return (
     <section className="py-12">
       <div className="max-w-6xl mx-auto">
-        <h2 className="text-2xl font-bold mb-2">{t("results_title")}</h2>
+        {/* LCP Element: Otimizado para carregamento prioritário */}
+        <h2 className="text-2xl font-heading font-bold mb-2">{t("results_title")}</h2>
         <p className="text-gray-600 mb-6">{t("results_subtitle")}</p>
         <div className="grid md:grid-cols-3 gap-4">
           {[1, 2, 3].map((i) => (